refactor(backend): core system improvements
- CLI: Enhanced adapter commands, diagnostics, profile management
- DB: Postgres improvements, domain adapter updates, process monitoring enhancements
- Orchestrator: Federation daemon improvements, code jobs, training updates
- Router: Calibration and orthogonal routing improvements
- Config: Loader improvements
- Crypto: Envelope and keychain provider updates
- Storage: Cleanup, monitoring, and quota improvements
- Telemetry: Metrics and unified events improvements
- System metrics: Anomaly detection, baselines, persistence improvements
- Error recovery: Corruption handling and validation improvements
- Git: Subsystem improvements
- Graph: Tensor improvements
- LoRA: Kernel, MLX FFI, lifecycle improvements
- Base LLM: Qwen improvements
- API: Type improvements
- Client: Native and UDS improvements
- Concurrent FS: Conflict resolution and locking improvements
- Policy: Hash watcher improvements
- Profiler: Scoring improvements
- Registry: Model improvements
- SecD: Enclave improvements
- Secure FS: Capability improvements
- Single-file adapter: MMAP loader and packager improvements
- Temp: Manager improvements
- Testing: Unified framework improvements

+++ crates/adapteros-system-metrics/src/telemetry.rs
@@ -64,7 +64,7 @@ impl SystemMetricsTelemetry {
         if metrics.cpu_usage > 90.0 {
             self.log_threshold_violation(&ThresholdViolationEvent::new(
                 "cpu_usage".to_string(),
-                metrics.cpu_usage,
+                metrics.cpu_usage as f32,
                 90.0,
                 "critical".to_string(),
             ))?;
@@ -73,7 +73,7 @@ impl SystemMetricsTelemetry {
         if metrics.memory_usage > 85.0 {
             self.log_threshold_violation(&ThresholdViolationEvent::new(
                 "memory_usage".to_string(),
-                metrics.memory_usage,
+                metrics.memory_usage as f32,
                 85.0,
                 "warning".to_string(),
             ))?;
